Moving EventSummary into shared code.
