+++ app/build.gradle
@@ -100,6 +100,7 @@ android {
 
 dependencies {
 
+    implementation project(":core-models")
     implementation "androidx.core:core-ktx:$rootProject.ext.versions.ktxCore"
     implementation "androidx.appcompat:appcompat:$rootProject.ext.versions.appCompat"
     implementation "com.google.android.material:material:$rootProject.ext.versions.material"
