+++ app/src/main/java/com/adammcneilly/pocketleague/event/data/EventService.kt
@@ -1,8 +1,8 @@
 package com.adammcneilly.pocketleague.event.data
 
 import com.adammcneilly.pocketleague.core.data.Result
+import com.adammcneilly.pocketleague.core.models.EventSummary
 import com.adammcneilly.pocketleague.eventoverview.domain.models.EventOverview
-import com.adammcneilly.pocketleague.eventsummary.domain.models.EventSummary
 import com.adammcneilly.pocketleague.swiss.domain.models.SwissStage
 
 /**
@@ -22,7 +22,7 @@ interface EventService {
      */
     suspend fun fetchUpcomingEvents(
         leagueSlug: String,
-    ): Result<List<EventSummary>>
+    ): Result<List<com.adammcneilly.pocketleague.core.models.EventSummary>>
 
     /**
      * Given an [eventId], fetch the overview information about this event.
