+++ app/src/main/java/com/adammcneilly/pocketleague/eventsummary/domain/usecases/FetchUpcomingEventsUseCase.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.pocketleague.eventsummary.domain.usecases
 
 import com.adammcneilly.pocketleague.core.data.Result
-import com.adammcneilly.pocketleague.eventsummary.domain.models.EventSummary
+import com.adammcneilly.pocketleague.core.models.EventSummary
 
 /**
  * Requests a list of upcoming RLCS events.
@@ -11,5 +11,5 @@ interface FetchUpcomingEventsUseCase {
     /**
      * @see [FetchUpcomingEventsUseCase]
      */
-    suspend operator fun invoke(): Result<List<EventSummary>>
+    suspend operator fun invoke(): Result<List<com.adammcneilly.pocketleague.core.models.EventSummary>>
 }
