+++ app/src/main/java/com/adammcneilly/pocketleague/eventsummary/domain/usecases/FetchUpcomingEventsUseCaseImpl.kt
@@ -1,8 +1,8 @@
 package com.adammcneilly.pocketleague.eventsummary.domain.usecases
 
 import com.adammcneilly.pocketleague.core.data.Result
+import com.adammcneilly.pocketleague.core.models.EventSummary
 import com.adammcneilly.pocketleague.event.data.EventService
-import com.adammcneilly.pocketleague.eventsummary.domain.models.EventSummary
 import javax.inject.Inject
 
 /**
@@ -12,7 +12,7 @@ class FetchUpcomingEventsUseCaseImpl @Inject constructor(
     private val service: EventService,
 ) : FetchUpcomingEventsUseCase {
 
-    override suspend fun invoke(): Result<List<EventSummary>> {
+    override suspend fun invoke(): Result<List<com.adammcneilly.pocketleague.core.models.EventSummary>> {
         val leagueSlug = "rlcs-2021-22-1"
 
         return service.fetchUpcomingEvents(
