+++ app/src/main/java/com/adammcneilly/pocketleague/eventsummary/ui/EventSummaryDisplayModel.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.pocketleague.eventsummary.ui
 
+import com.adammcneilly.pocketleague.core.models.EventSummary
 import com.adammcneilly.pocketleague.core.ui.UIImage
-import com.adammcneilly.pocketleague.eventsummary.domain.models.EventSummary
 
 /**
  * A user friendly summary of an [EventSummary] to be displayed on the UI.
