+++ core-models/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/models/EventSummary.kt
@@ -1,6 +1,4 @@
-package com.adammcneilly.pocketleague.eventsummary.domain.models
-
-import java.time.ZonedDateTime
+package com.adammcneilly.pocketleague.core.models
 
 /**
  * Represents an individual RLCS Event. Example params below.
@@ -10,7 +8,7 @@ import java.time.ZonedDateTime
  * @property[tournamentName] The broader name for the tournament being captured, such as
  * "RLCS 2021-22 Season - Winter Split Regional 1 - North America".
  * @property[tournamentImageUrl] An image URL to some image that will be used to identify this tournament.
- * @property[startDate] The [ZonedDateTime] that this event starts on.
+ * @property[startDateEpochSeconds] The time, in seconds since epoch, that this event begins.
  * @property[numEntrants] The number of teams registered or participating in this event. Optional if
  * no one has registered yet.
  * @property[isOnline] True if this is an event taking place digitally, false if in person LAN.
@@ -20,7 +18,7 @@ data class EventSummary(
     val eventName: String,
     val tournamentName: String,
     val tournamentImageUrl: String,
-    val startDate: ZonedDateTime,
+    val startDateEpochSeconds: Long,
     val numEntrants: Int?,
     val isOnline: Boolean,
 )
